fixup

+++ cpu/cc2538/include/periph_cpu.h
@@ -99,7 +99,7 @@ static inline uint8_t gpio_port_num(gpio_t pin)
  */
 static inline uint8_t gpio_pin_num(gpio_t pin)
 {
-    return (int)(pin & PIN_MASK);
+    return (uint8_t)(pin & PIN_MASK);
 }
 
 /**
@@ -123,7 +123,7 @@ static inline uint32_t gpio_pin_mask(gpio_t pin)
  */
 static inline uint8_t gpio_pp_num(gpio_t pin)
 {
-    return (gpio_port_num(pin) * 8) + gpio_pin_num(pin);
+    return (uint8_t)((gpio_port_num(pin) * 8) + gpio_pin_num(pin));
 }
 
 /**
